fix: update existing dependency and show a warning
If the package already existed, but in another kind of dependency, it updates that instead of potentially duplicating it another kind of dependencies property.

Run `npm-links yargs` and then again run `npm-links -D yargs`. Instead of now writing yargs in both "dependencies" and "devDependencies" it now only updates the existing yargs in "dependencies", and shows a warning.